+++ data_plumber/__init__.py
@@ -2,7 +2,8 @@ from .array import Pipearray
 from .error import PipelineError
 from .fork import Fork
 from .pipeline import Pipeline
-from .ref import Previous, First, Last, Next, Skip
+from .ref import PreviousN, Previous, First, NextN, Next, Skip, Last, \
+    StageById, StageByIndex, StageByIncrement
 from .stage import Stage
 
 __all__ = [
@@ -10,6 +11,7 @@ __all__ = [
     "PipelineError",
     "Fork",
     "Pipeline",
-    "First", "Previous", "Last", "Next", "Skip",
+    "PreviousN", "Previous", "First", "NextN", "Next", "Skip", "Last", \
+        "StageById", "StageByIndex", "StageByIncrement",
     "Stage",
 ]