+++ data_plumber/ref.py
@@ -5,22 +5,12 @@ This module defines classes for referencing and handling of flow-logic
 in a `Pipeline.run`.
 """
 
-from typing import Optional
+from typing import TypeAlias, Optional
 import abc
 from dataclasses import dataclass
 
 from .context import PipelineContext
-from .output import _StageRecord
-
-
-def _get_next_stageref(context: PipelineContext, id_: str) \
-        -> Optional[_StageRecord]:
-    """Helper to get most recent `_StageRecord` of 'future'-`Stage`."""
-    return next(
-        (stage for _, stage in enumerate(reversed(context.records))
-            if stage[0] == id_),
-        None
-    )
+from .error import PipelineError
 
 
 @dataclass
@@ -29,34 +19,35 @@ class StageRefOutput:
     Record class identifying a `Stage` during `Pipeline.run`.
 
     Properties:
-    stage -- string identifier of the references `Stage`
-             (required)
-    relative_index -- relative position in execution order
-                      * if positive, follows context.stages
-                      * if negative, follows context.records
-                      (required)
-    status -- most recent status value from the referenced `Stage`
-              (optional; `None` implies 'not executed yet')
+    stage -- string identifier of the referenced `Stage`
+    index -- absolte position in execution order (stages)
     """
     stage: str
-    relative_index: int
-    status: Optional[int] = None
+    index: int
 
 
-class StageRef(metaclass=abc.ABCMeta):
+class _StageRef(metaclass=abc.ABCMeta):
     """
     Base class enabling the definition of references to certain `Stage`s
     when executing a `Pipeline`. Only child-classes of this class are
     intended for explicit use.
     """
 
+    STAGEREF_ERROR_MSG = \
+        "Unable to resolve StageRef {target}{location} in Pipeline with " \
+        + "stages {stages}. Records until error: {records}"
+
     @staticmethod
+    def _format_records(records) -> str:
+        return ' > '.join(map(str, records))
+
+    @classmethod
     @abc.abstractmethod
-    def get(context: PipelineContext) -> Optional[StageRefOutput]:
+    def get(cls, context: PipelineContext) -> StageRefOutput:
         """
         Returns a `StageRefOutput` that given information on what is to
         be executed next. If this reference cannot be resolved within
-        the given context, the value `None` is returned.
+        the given context, a `PipelineError` is raised.
 
         Keyword arguments:
         context -- `Pipeline` execution context
@@ -64,74 +55,243 @@ class StageRef(metaclass=abc.ABCMeta):
         raise NotImplementedError("Missing definition of StageRef.get.")
 
 
-class Previous(StageRef):
+StageRef: TypeAlias = type[_StageRef]
+"""TypeAlias for type of _StageRef: type[_StageRef]"""
+
+
+def PreviousN(n: int, name: Optional[str] = None) -> StageRef:
+    """
+    Reference to a `Stage` that is `n` steps backwards in the
+    `Pipeline`'s list of `StageRecords`s.
+
+    Keyword arguments:
+    n -- number of steps
+    name -- used when generating an error-message
+            (default None; corresponds to 'Previous1', 'Previous2', ...)
+    """
+
+    if n < 0:
+        raise ValueError(
+            "'PreviousN'-StageRef only supports positive integers "
+            + f"(got '{n}')."
+        )
+
+    _name = name or f"Previous{n}"
+
+    class _(_StageRef):
+        @classmethod
+        def get(cls, context: PipelineContext) -> StageRefOutput:
+            if len(context.records) < n:
+                raise PipelineError(
+                    cls.STAGEREF_ERROR_MSG.format(
+                        target=_name,
+                        location=f" at index {str(context.current_position)}",
+                        stages=str(context.stages),
+                        records=cls._format_records(context.records)
+                    )
+                )
+            return StageRefOutput(
+                context.records[-1].id_,
+                context.records[-1].index
+            )
+    _.__doc__ = f"Reference to a `Stage` that is `{n}` steps backwards" \
+        + " in the `Pipeline`'s list of `StageRecords`s."
+
+    return _
+
+
+Previous = PreviousN(1, "Previous")
+"""Reference to the previous `Stage` during `Pipeline` execution."""
+Previous.__doc__ = \
     """Reference to the previous `Stage` during `Pipeline` execution."""
 
-    @staticmethod
-    def get(context: PipelineContext) -> Optional[StageRefOutput]:
-        if len(context.records) == 0:
-            return None
-        return StageRefOutput(
-            context.records[-1][0],
-            -1,
-            context.records[-1][2]
-        )
 
-
-class First(StageRef):
+class First(_StageRef):
     """Reference to the first `Stage` during `Pipeline` execution."""
 
-    @staticmethod
-    def get(context: PipelineContext) -> Optional[StageRefOutput]:
+    @classmethod
+    def get(cls, context: PipelineContext) -> StageRefOutput:
         if len(context.records) == 0:
-            return None
+            raise PipelineError(
+                cls.STAGEREF_ERROR_MSG.format(
+                    target="'First' ",
+                    location=f"at index {str(context.current_position)}",
+                    stages=str(context.stages),
+                    records=cls._format_records(context.records)
+                )
+            )
         return StageRefOutput(
-            context.records[0][0],
-            -context.current_position,
-            context.records[0][2]
+            context.records[0].id_,
+            context.records[0].index
         )
 
 
-class Last(StageRef):
+class Last(_StageRef):
     """Reference to the last `Stage` of registered `Stage`s in `Pipeline`."""
 
-    @staticmethod
-    def get(context: PipelineContext) -> Optional[StageRefOutput]:
+    @classmethod
+    def get(cls, context: PipelineContext) -> StageRefOutput:
+        if len(context.stages) == 0:
+            raise PipelineError(
+                cls.STAGEREF_ERROR_MSG.format(
+                    target="'Last' ",
+                    location="",
+                    stages=str(context.stages),
+                    records=cls._format_records(context.records)
+                )
+            )
         return StageRefOutput(
             context.stages[-1],
-            len(context.stages) - 1 - context.current_position,
-            x[2] if (x := _get_next_stageref(context, context.stages[-1]))
-                else None
+            len(context.stages) - 1
         )
 
 
-class Next(StageRef):
+def NextN(n: int) -> StageRef:
+    """
+    Reference to a `Stage` that is `n` steps forward in the `Pipeline`'s
+    list of `Stage`s.
+
+    Keyword arguments:
+    n -- number of steps
+    """
+
+    if n < 0:
+        raise ValueError(
+            f"'NextN'-StageRef only supports positive integers (got '{n}')."
+        )
+
+    class _(_StageRef):
+        @classmethod
+        def get(cls, context: PipelineContext) -> StageRefOutput:
+            stage_index = context.current_position + n
+            if context.loop:
+                stage_index = stage_index % len(context.stages)
+            try:
+                return StageRefOutput(
+                    context.stages[stage_index],
+                    stage_index
+                )
+            except IndexError:  # allow Pipeline to exit via StageRef
+                return StageRefOutput(
+                    "",
+                    stage_index
+                )
+    _.__doc__ = f"Reference to a `Stage` that is `{n}` steps forward" \
+        + " in the `Pipeline`'s list of `Stage`s."
+
+    return _
+
+
+Next = NextN(1)
+"""Reference to the next `Stage` of registered `Stage`s in `Pipeline`."""
+Next.__doc__ = \
     """Reference to the next `Stage` of registered `Stage`s in `Pipeline`."""
 
-    @staticmethod
-    def get(context: PipelineContext) -> Optional[StageRefOutput]:
-        if context.current_position + 1 >= len(context.stages):
-            return StageRefOutput("", 1, None)
-        identifier = context.stages[context.current_position + 1]
-        return StageRefOutput(
-            identifier,
-            1,
-            x[2] if (x := _get_next_stageref(context, identifier))
-                else None
-        )
-
-
-class Skip(StageRef):
+Skip = NextN(2)
+"""Reference to the `Stage` after next of registered `Stage`s in `Pipeline`."""
+Skip.__doc__ = \
     """Reference to the `Stage` after next of registered `Stage`s in `Pipeline`."""
 
-    @staticmethod
-    def get(context: PipelineContext) -> Optional[StageRefOutput]:
-        if context.current_position + 2 >= len(context.stages):
-            return StageRefOutput("", 2, None)
-        identifier = context.stages[context.current_position + 2]
-        return StageRefOutput(
-            identifier,
-            2,
-            x[2] if (x := _get_next_stageref(context, identifier))
-                else None
-        )
+
+def StageById(stage_id: str) -> StageRef:
+    """
+    Reference to a `Stage` with its id. (First occurrence in
+    `Pipeline`'s list of `Stage`s.)
+
+    Keyword arguments:
+    stage_id -- stage id
+    """
+
+    class _(_StageRef):
+        @classmethod
+        def get(cls, context: PipelineContext) -> StageRefOutput:
+            try:
+                index = context.stages.index(stage_id)
+            except ValueError as exc:
+                raise PipelineError(
+                    cls.STAGEREF_ERROR_MSG.format(
+                        target=f"to id '{stage_id}'",
+                        location="",
+                        stages=str(context.stages),
+                        records=cls._format_records(context.records)
+                    )
+                ) from exc
+            return StageRefOutput(
+                stage_id,
+                index
+            )
+    _.__doc__ = f"Reference to a `Stage` by the id '{stage_id}'. " \
+        + "(First occurrence in `Pipeline`'s list of `Stage`s.)"
+
+    return _
+
+
+def StageByIndex(stage_index: int) -> StageRef:
+    """
+    Reference to the `Stage` with an absolute index.
+
+    Keyword arguments:
+    stage_index -- absolute index
+    """
+
+    class _(_StageRef):
+        @classmethod
+        def get(cls, context: PipelineContext) -> StageRefOutput:
+            try:
+                stage_id = context.stages[stage_index]
+            except IndexError as exc:
+                raise PipelineError(
+                    cls.STAGEREF_ERROR_MSG.format(
+                        target=f"to index '{str(stage_index)}'",
+                        location="",
+                        stages=str(context.stages),
+                        records=cls._format_records(context.records)
+                    )
+                ) from exc
+            return StageRefOutput(
+                stage_id,
+                stage_index
+            )
+    _.__doc__ = f"Reference to a `Stage` by its index of {str(stage_index)}."
+
+    return _
+
+
+def StageByIncrement(index_increment: int) -> StageRef:
+    """
+    Reference to the `Stage` with a given relative index.
+
+    Keyword arguments:
+    index_increment -- relative index
+    """
+
+    class _(_StageRef):
+        @classmethod
+        def get(cls, context: PipelineContext) -> StageRefOutput:
+            stage_index = context.current_position + index_increment
+            try:
+                if context.loop:
+                    stage_index = stage_index % len(context.stages)
+                else:
+                    if stage_index < 0:
+                        raise IndexError(
+                            f"Bad index '{stage_index}' in StageRef of non-looping Pipeline."
+                        )
+                stage_id = context.stages[stage_index]
+            except IndexError as exc:
+                raise PipelineError(
+                    cls.STAGEREF_ERROR_MSG.format(
+                        target=f"to index '{str(stage_index)}'",
+                        location=f" (increment '{str(index_increment)}' from "
+                            + f"'{str(context.current_position)}')",
+                        stages=str(context.stages),
+                        records=cls._format_records(context.records)
+                    )
+                ) from exc
+            return StageRefOutput(
+                stage_id,
+                stage_index
+            )
+    _.__doc__ = f"Reference to a `Stage` by a relative index of {str(index_increment)}."
+
+    return _