+++ data_plumber/stage.py
@@ -8,7 +8,7 @@ a `Pipeline`.
 from typing import Optional, Callable, Any
 from uuid import uuid4
 
-from .ref import StageRef
+from .ref import StageRef, StageById, StageByIncrement
 
 
 class Stage:
@@ -26,6 +26,9 @@ class Stage:
     * `status` (output of `Stage.status`),
     * `count` (index of `Stage` in execution of `Pipeline`)
 
+    `Callable`s are executed in the order:
+    `primer` > `action` > `export` > `status` > `message`
+
     Example usage:
      >>> from data_plumber import Stage
      >>> Stage(
@@ -37,19 +40,29 @@ class Stage:
      <data_plumber.stage.Stage object at ...>
 
     Keyword arguments:
-    requires -- requirements for `Stage`-execution; dictionary with keys
-                being either `None`, a `StageRef`, or `str` (identifier
-                of a `Stage` in the context of a `Pipeline`; uses most
-                recent evaluation) and values being either an integer
-                (required output status of the keyed `Stage`) or a
-                `Callable` taking the status as an argument and
-                returning a `bool` (if it evaluates to `True`, the
-                `Stage`-requirement is met); `PipelineError` is raised
-                if references `Stage` has not yet been executed
+    requires -- requirements for `Stage`-execution being either `None`
+                or a dictionary with pairs of some reference to a `Stage`
+                and the required status (uses most recent evaluation);
+
+                key types are either `StageRef`, `str` (identifier of a
+                `Stage` in the context of a `Pipeline`), or `int`
+                (relative index in `Pipeline` stage arrangement);
+
+                values are either an integer value or a `Callable`
+                taking the status as an argument and returning a `bool`
+                (if it evaluates to `True`, the `Stage`-requirement is
+                met); `PipelineError` is raised if referenced `Stage`
+                has not yet been executed
     primer -- `Callable` for pre-processing data
               (kwargs: `out`, `count`)
               (default `lambda **kwargs: None`)
     action -- `Callable` for main-step of processing
+              (kwargs: `out`, `primer`, `count`)
+              (default `lambda **kwargs: None`)
+    export -- `Callable` that returns a dictionary of additional kwargs
+              to be exported to the parent `Pipeline`; in the following
+              `Stage`s, these kwargs are then available as if they were
+              provided with the `Pipeline.run`-command
               (kwargs: `out`, `primer`, `count`)
               (default `lambda **kwargs: None`)
     status -- `Callable` for generation of `Stage`'s integer exit status
@@ -63,16 +76,31 @@ class Stage:
     def __init__(
         self,
         requires: Optional[
-            dict[StageRef | str, int | Callable[[int], bool]]
+            dict[StageRef | str | int, int | Callable[[int], bool]]
         ] = None,
         primer: Callable[..., Any] = lambda **kwargs: None,
         action: Callable[..., Any] = lambda **kwargs: None,
+        export: Optional[Callable[..., Optional[dict[str, Any]]]] = None,
         status: Callable[..., int] = lambda **kwargs: 0,
         message: Callable[..., str] = lambda **kwargs: ""
     ) -> None:
-        self._requires = requires
+        if requires is None:
+            self._requires = None
+        else:
+            self._requires = {}
+            for k, v in requires.items():
+                if isinstance(k, str):
+                    self._requires[StageById(k)] = v
+                elif isinstance(k, int):
+                    self._requires[StageByIncrement(k)] = v
+                else:
+                    self._requires[k] = v
         self._primer = primer
         self._action = action
+        if export is None:
+            self._export: Callable[..., dict[str, Any]] = lambda **kwargs: {}
+        else:
+            self._export = export  # type: ignore[assignment]
         self._status = status
         self._message = message
         self._id = str(uuid4())
@@ -84,7 +112,7 @@ class Stage:
 
     @property
     def requires(self) -> Optional[
-        dict[StageRef | str, int | Callable[[int], bool]]
+        dict[StageRef, int | Callable[[int], bool]]
     ]:
         """Returns a `Stage`'s requirements."""
         return self._requires
@@ -98,6 +126,11 @@ class Stage:
     def action(self) -> Callable[..., Any]:
         """Returns a `Stage`'s `action` callable."""
         return self._action
+
+    @property
+    def export(self) -> Callable[..., Any]:
+        """Returns a `Stage`'s `export` callable."""
+        return self._export
 
     @property
     def status(self) -> Callable[..., int]: